Fix server animation ticks precision loss issue

+++ common/src/main/java/moe/plushie/armourers_workshop/core/armature/ArmaturePlugin.java
@@ -28,7 +28,7 @@ public abstract class ArmaturePlugin {
 
         float getPartialTicks();
 
-        float getAnimationTicks();
+        double getAnimationTicks();
 
         IPoseStack getPoseStack();
